image_search_mcp/requirements.txt
fastmcp>=2.13.0
chromadb
pillow
numpy
transformers
torch
requests

